Fixed return type

+++ src/Lock.php
@@ -203,6 +203,6 @@ final class Lock
      */
     protected function time()
     {
-        return round(microtime(true) * 1000);
+        return (int) round(microtime(true) * 1000);
     }
 }
